RAG/requirements.txt
# Core API & server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Client OpenAI-compatible (utilisé pour Ollama)
openai>=1.30.0

# Embeddings (multilingual, supports French)
sentence-transformers>=2.7.0

# Vector database
chromadb>=0.5.0

# Document parsers
pypdf>=3.0.0
python-docx>=0.8.11

# PostgreSQL (Supabase)
psycopg2-binary>=2.9.0
